# openapi/junban-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Junban API",
    "description": "Authenticated Axum server for Junban",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/reminders/events": {
      "get": {
        "tags": ["routes"],
        "operationId": "reminder_events",
        "responses": {
          "200": {
            "description": "Ephemeral reminder wake stream (not revisioned task events)",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderWakeEventDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": ["code", "message", "retryable"],
        "properties": {
          "code": {
            "type": "string"
          },
          "fields": {
            "type": ["object", "null"],
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": ["error", "request_id"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "type": "string"
          }
        }
      },
...
      "ReminderWakeEventDto": {
        "type": "object",
        "description": "Ephemeral wake payload for `/api/v1/reminders/events`.",
        "required": ["sequence", "server_now"],
        "properties": {
          "sequence": {
            "type": "integer",
            "format": "int64",
            "description": "Monotonic process-local sequence. Not a durable revision.",
            "minimum": 0
          },
          "server_now": {
            "type": "string",
            "format": "date-time",
            "description": "Server clock sample for the signal, RFC3339 UTC."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}